python/qt6/zoom-focus-demo/requirements.txt
imagingcontrol4>=1.2.0
imagingcontrol4pyside6
